+++ src/api/index.ts
@@ -1,13 +1,9 @@
 import axios from 'axios';
 
-import { API_KEY, API_URL } from 'constants/gen';
-
-const http = axios.create({
-  baseURL: API_URL
-});
+import { API_KEY } from 'constants/gen';
 
 export const httpGet = async (url: string) => {
-  return await http.get(url, {
+  return await axios.get(url, {
     headers: {
       Authorization: API_KEY
     }