+++ src/api/movies.ts
@@ -5,7 +5,7 @@ export const fetchMovies = async (search?: string) => {
   try {
     const queryString = search ?? search === '' ? `?q=${search}` : '';
     const response = await httpGet(`${API_URL as string}/movies${queryString}`);
-    return response.data;
+    return response?.data;
   } catch (error) {
     console.log('error is ', error);
     throw error;