+++ src/hooks/useMoviesSearch.ts
@@ -1,15 +1,17 @@
 import { useEffect, useState } from 'react';
 
+import { useDebounce } from 'hooks/useDebounce';
 import { useStore } from 'store/movies';
 
 const useMoviesSearch = () => {
   const [searchQuery, setSearchQuery] = useState('');
+  const debouncedSearchQuery = useDebounce(searchQuery, 500);
   const { searchedMovies, searchLoadingState, searchMoviesByTitle } =
     useStore();
 
   useEffect(() => {
-    searchMoviesByTitle(searchQuery);
-  }, [searchMoviesByTitle, searchQuery]);
+    searchMoviesByTitle(debouncedSearchQuery);
+  }, [searchMoviesByTitle, debouncedSearchQuery]);
 
   return {
     searchedMovies,