+++ src/testutils/mockData.ts
@@ -1,0 +1,4 @@
+export const getMockMovie = () => ({
+  id: 'movie_id_1',
+  title: 'Movie 1'
+});